Correção namespace da Dimob e mesclagem dos enums com as constantes
