+++ FiscalBr/FiscalBr.Dimob/DimobT9.cs
@@ -1,4 +1,6 @@
-﻿namespace FiscalBr.Dimob
+﻿using FiscalBr.Common.Dimob;
+
+namespace FiscalBr.Dimob
 {
     /// <summary>
     /// Trailler da Declaração - T9
